Organize files, add context/reducer/placeholders
