--- src/App.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1 className="App-logo">🕵️</h1>
-        <p>
-          Welcome to KidDo
-        </p>
-      </header>
-    </div>
-  );
-}
-
-export default App;
