+++ src/common/modules/appReducer.js
@@ -1,0 +1,9 @@
+export const appReducer = (state, action) => {
+  switch(action.type) {
+    case 'SOME_CASE':
+      // DO THINGS
+      return {...state}
+    default:
+      return state;
+  }
+}
