+++ src/setup/service-worker.js
